// src/app/hoteles/models/hoteles.model.ts
export class Hoteles {
    nit: string;
    nombre: string;
    direccion: string;
    telefono: number;
    descripcion: string;
    imagen: string;
    categoria: string;
    ciudad: string;
    precioMasBajo: number;
 }

 export class BusquedaFormulario {
    ciudad: string;
    categoria: string;
    precioDe: string;
    precioHasta: string;
 }